refactor(system): share platform-independent account config
