+++ src/system/accounts_non_linux.rs
@@ -1,49 +1,11 @@
 use std::collections::BTreeSet;
-use std::path::PathBuf;
 
 use eyre::{Result, bail};
-use serde::Deserialize;
 
 use crate::config::Config;
 use crate::system::resources::{ResourceAction, ResourceId, ResourcePlan};
 
-#[derive(Clone, Copy, Debug, Default, Deserialize, Eq, PartialEq)]
-#[serde(rename_all = "snake_case")]
-pub(crate) enum AccountState {
-    #[default]
-    Present,
-    Absent,
-}
-
-#[derive(Clone, Debug, Default, Deserialize)]
-pub(crate) struct GroupTomlConfig {
-    #[serde(default)]
-    pub state: AccountState,
-    pub gid: Option<u32>,
-    #[serde(default)]
-    pub system: bool,
-}
-
-#[derive(Clone, Debug, Default, Deserialize)]
-pub(crate) struct UserTomlConfig {
-    #[serde(default)]
-    pub state: AccountState,
-    pub uid: Option<u32>,
-    pub group: Option<String>,
-    pub groups: Option<Vec<String>>,
-    #[serde(default)]
-    pub exclusive_groups: bool,
-    pub home: Option<PathBuf>,
-    pub shell: Option<PathBuf>,
-    pub comment: Option<String>,
-    #[serde(default)]
-    pub system: bool,
-    pub create_home: Option<bool>,
-    #[serde(default)]
-    pub move_home: bool,
-    #[serde(default)]
-    pub remove_home: bool,
-}
+pub(crate) use super::accounts_common::*;
 
 #[derive(Clone, Debug)]
 pub(crate) struct GroupRequest {
